Move helper methods to SlackHelper class
