+++ tests/Adapters/Microsoft.Bot.Builder.Adapters.Slack.TestBot/appsettings.json
@@ -1,0 +1,6 @@
+{
+  "MicrosoftAppId": "",
+  "MicrosoftAppPassword": "",
+  "VerificationToken": "",
+  "BotToken": ""
+}
